fix various bugs

+++ app/ca/teams/views.py
@@ -58,8 +58,12 @@ def edit_team(team_id):
 @login_required
 def delete_team(team_id):
     team = Team.query.get_or_404(team_id)
-    db.session.delete(team)
-    db.session.commit()
+    if len(team.members) or len(team.treasury.transactions) or len(team.activities) or len(team.announcements) or len(team.files):
+        flash('Cannot delete as there are entities associated with this team, like members, activities, announcements, files or transactions')
+    else:
+        db.session.delete(team)
+        db.session.delete(team.treasury)
+        db.session.commit()
+        flash('You have successfully deleted the team.')
 
-    flash('You have successfully deleted the team.')
     return redirect(url_for('ca.list_teams'))
